+++ app/src/main/java/com/ceaver/assin/alerts/AlertWorker.kt
@@ -5,6 +5,7 @@ import androidx.work.Worker
 import androidx.work.WorkerParameters
 import com.ceaver.assin.logging.LogRepository
 import com.ceaver.assin.markets.TitleRepository
+import java.math.BigDecimal
 
 class AlertWorker(appContext: Context, workerParams: WorkerParameters) : Worker(appContext, workerParams) {
     override fun doWork(): Result {
@@ -15,7 +16,7 @@ class AlertWorker(appContext: Context, workerParams: WorkerParameters) : Worker(
     private fun checkAlert(alert: Alert) {
         val price = TitleRepository.lookupPrice(alert.symbol, alert.reference)
         if(price.isPresent) {
-            val currentPrice = price.get()
+            val currentPrice = BigDecimal.valueOf(price.get())
             val result = alert.alertType.check(alert, currentPrice)
             result.ifPresent { AlertRepository.updateAlert(it); checkAlert(it); AlertNotification.notify(alert.symbol, alert.reference, targetPrice(alert, currentPrice), currentPrice) }
         } else {
@@ -23,7 +24,7 @@ class AlertWorker(appContext: Context, workerParams: WorkerParameters) : Worker(
         }
     }
 
-    private fun targetPrice(alert: Alert, currentPrice: Double): Double {
+    private fun targetPrice(alert: Alert, currentPrice: BigDecimal): BigDecimal {
         return when {
             currentPrice <= (alert.source - alert.target) -> alert.source - alert.target
             currentPrice >= (alert.source + alert.target) -> alert.source + alert.target
